Cinder icehouse-rc1 milestone (2014.1.rc1)

+++ cinder/volume/drivers/san/hp/hp_3par_common.py
@@ -120,10 +120,12 @@ class HP3PARCommon(object):
         2.0.4 - Allow volumes created from snapshots to be larger bug #1279478
         2.0.5 - Fix extend volume units bug #1284368
         2.0.6 - use loopingcall.wait instead of time.sleep
+        2.0.7 - Allow extend volume based on snapshot bug #1285906
+        2.0.8 - Fix detach issue for multiple hosts Bug #1288927
 
     """
 
-    VERSION = "2.0.6"
+    VERSION = "2.0.8"
 
     stats = {}
 
@@ -248,14 +250,39 @@ class HP3PARCommon(object):
         volume_name = self._get_3par_vol_name(volume['id'])
         old_size = volume['size']
         growth_size = int(new_size) - old_size
-        LOG.debug("Extending Volume %s from %s to %s, by %s GB." %
-                  (volume_name, old_size, new_size, growth_size))
+        LOG.debug(_("Extending Volume %(vol)s from %(old)s to %(new)s, "
+                    " by %(diff)s GB.") %
+                  {'vol': volume_name, 'old': old_size, 'new': new_size,
+                   'diff': growth_size})
         growth_size_mib = growth_size * units.KiB
+        self._extend_volume(volume, volume_name, growth_size_mib)
+
+    def _extend_volume(self, volume, volume_name, growth_size_mib,
+                       _convert_to_base=False):
         try:
+            if _convert_to_base:
+                LOG.debug(_("Converting to base volume prior to growing."))
+                self._convert_to_base_volume(volume)
             self.client.growVolume(volume_name, growth_size_mib)
-        except Exception:
-            with excutils.save_and_reraise_exception():
-                LOG.error(_("Error extending volume %s") % volume)
+        except Exception as ex:
+            with excutils.save_and_reraise_exception() as ex_ctxt:
+                if (not _convert_to_base and
+                    isinstance(ex, hpexceptions.HTTPForbidden) and
+                        ex.get_code() == 150):
+                        # Error code 150 means 'invalid operation: Cannot grow
+                        # this type of volume'.
+                        # Suppress raising this exception because we can
+                        # resolve it by converting it into a base volume.
+                        # Afterwards, extending the volume should succeed, or
+                        # fail with a different exception/error code.
+                        ex_ctxt.reraise = False
+                        self._extend_volume(volume, volume_name,
+                                            growth_size_mib,
+                                            _convert_to_base=True)
+                else:
+                    LOG.error(_("Error extending volume: %(vol)s. "
+                                "Exception: %(ex)s") %
+                              {'vol': volume_name, 'ex': ex})
 
     def _get_3par_vol_name(self, volume_id):
         """Get converted 3PAR volume name.
@@ -424,6 +451,21 @@ class HP3PARCommon(object):
                                        'hp3par_cpg')})
         self.stats = stats
 
+    def _get_vlun(self, volume_name, hostname):
+        """find a VLUN on a 3PAR host."""
+        vluns = self.client.getHostVLUNs(hostname)
+        found_vlun = None
+        for vlun in vluns:
+            if volume_name in vlun['volumeName']:
+                found_vlun = vlun
+                break
+
+        msg = (_("3PAR vlun %(name)s not found on host %(host)s") %
+               {'name': volume_name, 'host': hostname})
+        if found_vlun is None:
+            LOG.warn(msg)
+        return found_vlun
+
     def create_vlun(self, volume, host, nsp=None):
         """Create a VLUN.
 
@@ -431,17 +473,19 @@ class HP3PARCommon(object):
         """
         volume_name = self._get_3par_vol_name(volume['id'])
         self._create_3par_vlun(volume_name, host['name'], nsp)
-        return self.client.getVLUN(volume_name)
+        return self._get_vlun(volume_name, host['name'])
 
     def delete_vlun(self, volume, hostname):
         volume_name = self._get_3par_vol_name(volume['id'])
-        vlun = self.client.getVLUN(volume_name)
-        # VLUN Type of MATCHED_SET 4 requires the port to be provided
-        if self.VLUN_TYPE_MATCHED_SET == vlun['type']:
-            self.client.deleteVLUN(volume_name, vlun['lun'], hostname,
-                                   vlun['portPos'])
-        else:
-            self.client.deleteVLUN(volume_name, vlun['lun'], hostname)
+        vlun = self._get_vlun(volume_name, hostname)
+
+        if vlun is not None:
+            # VLUN Type of MATCHED_SET 4 requires the port to be provided
+            if self.VLUN_TYPE_MATCHED_SET == vlun['type']:
+                self.client.deleteVLUN(volume_name, vlun['lun'], hostname,
+                                       vlun['portPos'])
+            else:
+                self.client.deleteVLUN(volume_name, vlun['lun'], hostname)
 
         try:
             self._delete_3par_host(hostname)
